rainbow/requirements.txt
django
djangorestframework
python-dotenv
djoser
drf-spectacular
psycopg2-binary
django-extensions
pytest
model_bakery
faker
qrcode[pil]
django_otp
gunicorn
django_widget_tweaks